Post-submission bug fix: add implicit constraints
`subset_accepting` did not enforce constraints 0 < x[i] < 1.

These constraints are maintained implicitly when none are binding, under uniform sampling, and by the distance optimization.

VI can generate points outside of these constraints, so long as they are valid under the explicit constraints. In the example problems, it appears that the hypercube surface is non-binding for problem instances where VI is selected, so the bug impacts none of the examples.